Practica DAW

+++ practica77/src/practica77/Practica77.java
@@ -25,7 +25,7 @@ public class Practica77 {
         
         //Declarar radio
         System.out.println("Radio:");
-        op.setRadio(teclado.nextInt());
+        op.setRadio(teclado.nextFloat());
         
         //Operacion Longitud
         op.longitud(op.getRadio());
